+++ src/main/java/com/ramirocantu/smithman/ShutdownTask.java
@@ -1,13 +1,13 @@
-package com.targren.forgeautoshutdown;
+package com.ramirocantu.smithman;
 
+import com.ramirocantu.smithman.util.Chat;
 import net.minecraft.server.MinecraftServer;
 import net.minecraftforge.fml.common.FMLCommonHandler;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
 import org.apache.logging.log4j.Logger;
-import com.targren.forgeautoshutdown.util.Chat;
-import com.targren.forgeautoshutdown.util.Server;
+import com.ramirocantu.smithman.util.Server;
 
 import java.text.Format;
 import java.text.SimpleDateFormat;
